pyproject.toml
[project]
name = "minemeets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "langchain_ollama",
    "numpy>=2.3.4",
    "ollama>=0.6.0",
    "openai-whisper>=20250625",
    "pinecone-client==2.2.4",
    "pydub>=0.25.1",
    "python-dotenv>=1.1.1",
    "requests",
    "sentence-transformers>=5.1.1",
    "streamlit>=1.50.0",
    "tqdm>=4.67.1",
]
